Unify `also_with_wasmfs` decorator. NFC
This also means that also_with_wasmfs in test_core.py now used
parameterization.
